refactor contents module
