+++ src/plugin/core/contents/__tests__/fixtures/invalid-maxFileSize/invalid-maxFileSize-contents.zip
