+++ src/plugin/core/contents/__tests__/fixtures/plugin-manifest-v1/image/icon.png
