+++ src/plugin/core/contents/__tests__/fixtures/plugin-manifest-v2/manifest.json
