+++ src/plugin/core/plugin/__tests__/fixtures/contents.zip
